API

Blobs

The detect_blobs function processes an image to identify relevant features in the form of blobs, which are represented by the Blob type.

ParticleTracking.AbstractBlob — Type

AbstractBlob{T,S,N}

Abstract interface for blob types

ParticleTracking.Blob — Type

struct Blob{T,S,N} <: AbstractBlob{T,S,N}
    location::CartesianIndex{N}
    σ::S
    amplitude::T
    m0::T
    m2::T
    intensity_map
end

The properties of a Blob can be accessed through the following functions.

ParticleTracking.location — Function

location(blob)

Return the estimated location of the blob including subpixel refinement. The raw location can be obtained with location_raw.

Missing docstring.

Missing docstring for location_raw. Check Documenter's build log for details.

ParticleTracking.scale — Function

scale(blob)

Return the estimated scale of the blob along each dimension.

ParticleTracking.amplitude — Function

amplitude(blob)

Return the peak amplitude of the blob.

ParticleTracking.radius — Function

radius(blob)

Return the equivalent spherical radius of the blob.

ParticleTracking.zeroth_moment — Function

zeroth_moment(blob)

Return the zeroth intensity moment of the blob image.

ParticleTracking.second_moment — Function

second_moment(blob)

Return the second intensity moment of the blob image.

ParticleTracking.intensity_map — Function

intensity_map(blob)

Return the image of the blob.

Cost functions

blobtracking links blob into trajectories by finding the optimal solution to the minimum cost flow problem defined by the keyword cost function. Two cost functions are provided out of the box:

ParticleTracking.QuadraticCost — Function

QuadraticCost(A, B; g0=1.0, g2=1.0)

Evaluate the cost of linking two blobs A and B using the Euclidean (𝐿²) norm.

The cost includes the spatial distance of the two blobs, and the distance in their zeroth and second intensity moments.

Keywords

g0::Real = 1.0: weight factor for the zeroth moment

g2::Real = 1.0: weight factor for the second moment

ParticleTracking.PCost — Function

PCost(A, B; p=1, g0=1.0, g2=1.0)

Evaluate the cost of linking two blobs A and B using the 𝐿ᵖ norm.

The cost includes the spatial distance of the two blobs, and the distance in their zeroth and second intensity moments.

Keywords

p::Real = 1: the norm exponent; p=1 corresponds to the Manhattan metric, p=Inf to the Chebyshev metric, and p=2 to the Euclidean metric (in which case QuadraticCost provides slightly better performance)

g0::Real = 1.0: weight factor for the zeroth moment

g2::Real = 1.0: weight factor for the second moment

However, arbitrary cost functions can be defined by the user, and they will "just work" as long as they respect the API. A cost function must:

take two and only two positional arguments which are subtypes of AbstractBlob

accept any other relevant argument in the form of a keyword argument

not modify the state of the blobs

always return a non-negative real number

Cost functions do not have to explicitly take care of maximum allowed distances, they should only evaluate the cost of linking the two input blobs.

Trajectories

The result of the blobtracking is a vector of Trajectory objects.

Missing docstring.

Missing docstring for Trajectory. Check Documenter's build log for details.

A Trajectory contains a vector of blobs and an associated vector of time points, which are not necessarily contiguous (in fact, when tracking with memory>1, we allow a trajectory to contain gaps).

All the functions of the Blob API can be also called on Trajectory objects, and, they will return a vector with the value of the requested property at each step along the trajectory.

A Trajectory also supports the following functions

Base.length — Function

length(trajectory)

Return the length of the trajectory.

ParticleTracking.timestamps — Function

timestamps(trajectory)

Return the times along the trajectory at which blobs have been measured.

ParticleTracking.startpoint — Function

startpoint(trajectory)

Return the time at which the trajectory started.

ParticleTracking.endpoint — Function

endpoint(trajectory)

Return the time at which the trajectory ended.